Did some quick performance testing of integration vs functional layers.
Found the run times of these 1000 test to be the same both as reported
by the testrunner and as reported by the `time` command.  Given this,
I think we should just standardize on functional testing since it has
no apparent additional cost and is much less surprising if a user
unknowingly calls `transaction.commit`.